refactor(signal/scan.py) move plot to submodule
Move plot class and associated functions to own submodule in order to
reduce the amount of code held in signal/scan.py. Generally more logical
to have them separated for clarity. Does not close issue #33 but is a
good start.

Addresses issue #33
